router refactor
