+++ src/server/root.ts
@@ -1,4 +1,3 @@
-import { algorithmRouter } from "./routers/algorithm";
 import { institutionRouter } from "./routers/institution";
 import { projectRouter } from "./routers/project";
 import { userRouter } from "./routers/user";
@@ -8,7 +7,6 @@ export const appRouter = createTRPCRouter({
   project: projectRouter,
   user: userRouter,
   institution: institutionRouter,
-  algorithm: algorithmRouter,
 });
 
 export type AppRouter = typeof appRouter;
