+++ src/server/routers/institution/instance/index.ts
@@ -1,0 +1,286 @@
+import { Role, Stage } from "@prisma/client";
+import { z } from "zod";
+
+import { instanceParamsSchema } from "@/lib/validations/params";
+import {
+  adminProcedure,
+  createTRPCRouter,
+  protectedProcedure,
+  stageAwareProcedure,
+} from "@/server/trpc";
+import { algorithmRouter } from "./algorithm";
+import { matchingRouter } from "./matching";
+
+export const instanceRouter = createTRPCRouter({
+  matching: matchingRouter,
+  algorithm: algorithmRouter,
+
+  access: stageAwareProcedure
+    .input(z.object({ params: instanceParamsSchema }))
+    .query(
+      async ({
+        ctx,
+        input: {
+          params: { group, subGroup, instance },
+        },
+      }) => {
+        const user = ctx.session.user;
+
+        const adminInSpace = await ctx.db.adminInSpace.findFirst({
+          where: {
+            allocationGroupId: group,
+            allocationSubGroupId: subGroup,
+            userId: user.id,
+          },
+        });
+        if (adminInSpace) return true;
+
+        const stage = ctx.stage;
+        const { role } = await ctx.db.userInInstance.findFirstOrThrow({
+          where: {
+            allocationGroupId: group,
+            allocationSubGroupId: subGroup,
+            allocationInstanceId: instance,
+            userId: user.id,
+          },
+          select: { role: true },
+        });
+
+        const supervisorStages: Stage[] = [Stage.SETUP];
+        if (role === Role.SUPERVISOR) return !supervisorStages.includes(stage);
+
+        const studentStages: Stage[] = [Stage.SETUP, Stage.PROJECT_SUBMISSION];
+        if (role === Role.STUDENT) return !studentStages.includes(stage);
+
+        return true;
+      },
+    ),
+
+  // ! not in use
+  // algorithms: adminProcedure
+  //   .input(
+  //     z.object({
+  //       params: instanceParamsSchema,
+  //     }),
+  //   )
+  //   .query(
+  //     async ({
+  //       ctx,
+  //       input: {
+  //         params: { group, subGroup, instance },
+  //       },
+  //     }) => {
+  //       const data = await ctx.db.algorithm.findMany({
+  //         where: {
+  //           allocationGroupId: group,
+  //           allocationSubGroupId: subGroup,
+  //           allocationInstanceId: instance,
+  //           NOT: builtInAlgSchema.options.map((algName) => ({ algName })),
+  //         },
+  //         select: {
+  //           algName: true,
+  //           displayName: true,
+  //           description: true,
+  //           flag1: true,
+  //           flag2: true,
+  //           flag3: true,
+  //         },
+  //       });
+  //       return data;
+  //     },
+  //   ),
+
+  currentStage: protectedProcedure
+    .input(z.object({ params: instanceParamsSchema }))
+    .query(
+      async ({
+        ctx,
+        input: {
+          params: { group, subGroup, instance },
+        },
+      }) => {
+        const { stage } = await ctx.db.allocationInstance.findFirstOrThrow({
+          where: {
+            allocationGroupId: group,
+            allocationSubGroupId: subGroup,
+            id: instance,
+          },
+          select: {
+            stage: true,
+          },
+        });
+        return stage;
+      },
+    ),
+
+  setStage: adminProcedure
+    .input(
+      z.object({
+        params: instanceParamsSchema,
+        stage: z.nativeEnum(Stage),
+      }),
+    )
+    .mutation(
+      async ({
+        ctx,
+        input: {
+          params: { group, subGroup, instance },
+          stage,
+        },
+      }) => {
+        await ctx.db.allocationInstance.update({
+          where: {
+            allocationGroupId_allocationSubGroupId_id: {
+              allocationGroupId: group,
+              allocationSubGroupId: subGroup,
+              id: instance,
+            },
+          },
+          data: { stage },
+        });
+      },
+    ),
+
+  projectAllocations: adminProcedure
+    .input(z.object({ params: instanceParamsSchema }))
+    .query(
+      async ({
+        ctx,
+        input: {
+          params: { group, subGroup, instance },
+        },
+      }) => {
+        const byStudent = await ctx.db.projectAllocation.findMany({
+          where: {
+            allocationGroupId: group,
+            allocationSubGroupId: subGroup,
+            allocationInstanceId: instance,
+          },
+          // ! currently selects all users not just students
+          select: {
+            student: {
+              select: {
+                user: {
+                  select: { id: true, name: true, email: true },
+                },
+              },
+            },
+            project: {
+              select: {
+                id: true,
+                supervisor: { select: { user: { select: { name: true } } } },
+              },
+            },
+            studentRanking: true,
+          },
+        });
+
+        const byProject = await ctx.db.projectAllocation.findMany({
+          where: {
+            allocationGroupId: group,
+            allocationSubGroupId: subGroup,
+            allocationInstanceId: instance,
+          },
+          select: {
+            project: {
+              select: {
+                id: true,
+                title: true,
+                capacityLowerBound: true,
+                capacityUpperBound: true,
+                supervisor: {
+                  select: { user: { select: { id: true, name: true } } },
+                },
+              },
+            },
+            userId: true,
+            studentRanking: true,
+          },
+        });
+
+        const bySupervisor = await ctx.db.projectAllocation.findMany({
+          where: {
+            allocationGroupId: group,
+            allocationSubGroupId: subGroup,
+            allocationInstanceId: instance,
+          },
+          select: {
+            project: {
+              select: {
+                id: true,
+                title: true,
+                supervisor: {
+                  select: {
+                    user: { select: { id: true, name: true, email: true } },
+                    supervisorInstanceDetails: {
+                      where: {
+                        allocationGroupId: group,
+                        allocationSubGroupId: subGroup,
+                        allocationInstanceId: instance,
+                      },
+                      select: {
+                        projectAllocationLowerBound: true,
+                        projectAllocationTarget: true,
+                        projectAllocationUpperBound: true,
+                      },
+                      take: 1, // * does not do anything
+                    },
+                  },
+                },
+              },
+            },
+            userId: true,
+            studentRanking: true,
+          },
+        });
+
+        return { byStudent, byProject, bySupervisor };
+      },
+    ),
+
+  supervisors: protectedProcedure
+    .input(z.object({ params: instanceParamsSchema }))
+    .query(
+      async ({
+        ctx,
+        input: {
+          params: { group, subGroup, instance },
+        },
+      }) => {
+        return await ctx.db.userInInstance.findMany({
+          where: {
+            allocationGroupId: group,
+            allocationSubGroupId: subGroup,
+            allocationInstanceId: instance,
+            role: Role.SUPERVISOR,
+          },
+          select: {
+            user: { select: { id: true, name: true, email: true } },
+          },
+        });
+      },
+    ),
+
+  students: protectedProcedure
+    .input(z.object({ params: instanceParamsSchema }))
+    .query(
+      async ({
+        ctx,
+        input: {
+          params: { group, subGroup, instance },
+        },
+      }) => {
+        return await ctx.db.userInInstance.findMany({
+          where: {
+            allocationGroupId: group,
+            allocationSubGroupId: subGroup,
+            allocationInstanceId: instance,
+            role: Role.STUDENT,
+          },
+          select: {
+            user: { select: { id: true, name: true, email: true } },
+          },
+        });
+      },
+    ),
+});
